% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_breakpoints}
\alias{get_breakpoints}
\title{Get the breakpoint locations for the given x and y values, and given parameters}
\usage{
get_breakpoints(x, y, n_bp, fit_type, min_len, ssr_mat)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{fit_type}{type of equation to fit (linear, exponential, or quadratic)}

\item{min_len}{the minimum number of x values between breakpoints}

\item{ssr_mat}{the sum squared residuals matrix}

\item{n_bps}{number of breakpoints to estimate}
}
\value{
a named list where bps is the breakpoint indices and ssr is a list of ssr values for each segment
}
\description{
Get the breakpoint locations for the given x and y values, and given parameters
}
