% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piecewise_regression}
\alias{piecewise_regression}
\title{Estimate the equation parameters and residuals for each piece of the data}
\usage{
piecewise_regression(x, y, bps, fit_type)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{bps}{breakpoint indices}

\item{fit_type}{type of equation to fit (linear, exponential, or quadratic)}
}
\value{
a named list of the intercept, coefficients, sum squared residuals, and residual values for each segment
}
\description{
Estimate the equation parameters and residuals for each piece of the data
}
