% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{precompute_ssr_mat}
\alias{precompute_ssr_mat}
\title{Compute the sum squared residuals matrix}
\usage{
precompute_ssr_mat(x, y, fit_type)
}
\arguments{
\item{x}{x values}

\item{y}{yvalues}

\item{fit_type}{the type of equation to fit (linear, exponential, or quadratic)}
}
\value{
sum squared residual matrix
}
\description{
Compute the sum squared residuals matrix
}
