% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataFinder.R
\docType{class}
\name{StrataFinder-class}
\alias{StrataFinder-class}
\alias{StrataFinder}
\title{Precompute the sum squared residuals matrix}
\arguments{
\item{n}{number of breakpoings}

\item{min_bps}{the minimum number of breakpoints to try}

\item{max_bps}{the maximum number of breakpoints to try}

\item{early_stop}{TRUE to exit once increasing the number of breakpoints does not lead to significantly different residuals. FALSE to run all possible breakpoint locations.}

\item{quiet}{TRUE to not print any updates, FALSE to print updates}

\item{residuals1}{the first set of residuals to compare}

\item{residuals2}{the second set of residuals to compare}

\item{test}{statistical test type to compare residual distributions (t, f, either, or both)}

\item{show_error}{TRUE to show sum squared residual error, FALSE to just show breakpoints}
}
\value{
the sum squared residual matrix
Find a specified number of strata

a named list with the breakpoints and their indices
Find the number of breakpoints and their location that best fit the data

a named list with the best number of breakpoints, their locations, and their indices
Perform a statistical comparison between the two sets of residuals

the p-value of the statistical test (in the case of "either, this is the lower p-value of the two tests, in the case of "both", this is the higher p-value of the two tests)
Get the error range of each breakpoint

named list with the bounds and SSR values for each breakpoint
Run piecewise regression over the data

a named list of the intercept, coefficients, sum squared residuals, and residual values for each segment
Run piecewise regression over the data

a named list of the intercept, coefficients, sum squared residuals, and residual values for each segment
Plot the data with the current saved breakpoints and fit type

a ggplot2 plot object
}
\description{
Precompute the sum squared residuals matrix
}

